feat: added content base settings modal
